requirements.txt
git+https://github.com/craftoid/twitterbot
numpy
pillow
